+++ agent/cmd/labwatch/main.go
@@ -22,7 +22,7 @@ import (
 )
 
 var (
-	version   = "0.3.2"
+	version   = "0.3.3"
 	buildDate = "dev"
 )